Found issue on User constructor; added password salt to DB
The issue was with the getUserAttribute method in the DBManager. I may fix this or may give it a complete overhaul and split it into multiple methods. Either way, authentication is broken as I was generating a new salt every time and wondering why they weren't the same (duh, I just made a new random one). Will store the salt with the password in the DB and hopefully avoid this from here on in.